Fix regression. Read more...
Regression introduced in a bug fix (prev commit) were pressing
the send to server email button would not allow for answers to be
changes.

That regression introduced when past answers were changed but the
subsequent questions remained visible when they had to be removed.

This regression fix fixes the above problem and still ensures the
send to servey email button bug is also fixed.